// src/interfaces/Owners.tsx
// Generated by https://quicktype.io

export interface AllOwnersResponse {
      owners: Owner[];
}

export interface OwnerResponse {
      owner: Owner;
}

export interface Owner {
      _id: string;
      name: string;
      lastName: string;
      email: string;
      phoneNumber: string;
      createdAt: string;
      updatedAt: string;
}