// src/interfaces/Players.tsx
// Generated by https://quicktype.io

export interface Player {
  _id: string;
  name: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  createdAt: string;
  updatedAt: string;
}

export interface PlayersResponse {
  players: Player[];
}

export interface PlayerResponse {
  player: Player;
}

export interface PostPlayerResponse {
  newPlayer: Player;
}

export interface LoginPlayerResponse {
  status: string;
  token: string;
  player: Player;
}